test/instanceha/requirements.txt
PyYAML>=5.1
requests>=2.20.0
